requirements.txt
PyQt6
pyjson
pyyaml
tabulate
psycopg2-binary